Support Zulu bundle flags
Some Zulu distributions come with extra flags. I care about EA builds,
but there are other flags, which currently fail the string.split test.

This PR attempts to handle these flags on the Zulu part of the version
string. The Java version string can also contain -ea, but I assume the
'rest' of the string is the Java version string.